+++ NEWS.md
@@ -4,6 +4,7 @@
 - improve coloring of barcodes in bubble plots
 - update plotDetectedBarcodes
 - collapseReplicates and thresholdCounts update the lib.size column to the sum of the merged replicates or filtered sample.
+- fix bug where small dots were plotted for barcodes that were not present in sample in plotOrderedBubble and plotBarcodeBubble
 
 # bartools 1.0.0