+++ NEWS.md
@@ -1,4 +1,4 @@
-# bartools 1.0.1
+# bartools 1.1.0
 
 - Fix plotDetectedBarcodes and calcPercentileBarcodes to be inclusive, i.e. at least 1 barcode makes up top x percentile of a sample
 - improve coloring of barcodes in bubble plots